// src/types/ISAScoutForm.ts
interface ISAScoutForm {
  eventKey: string;
  matchLevel: string;
  matchNumber: number;
  teamNumber: number;
  alliance: string;
  robotNumber: number;
  deviceTeamNumber: number;
  deviceId: string;
  scoutTeamNumber: number;
  scoutName: string;
  flag: string;
  noShow: number;
  startingLocationA: number;
  startingLocationB: number;
  startingLocationC: number;
  died: number;
  playedDefense: number;
  goodAtCoralL1: number;
  goodAtCoralL2: number;
  goodAtCoralL3: number;
  goodAtCoralL4: number;
  goodAtAlgaeNet: number;
  goodAtAlgaeProcessor: number;
  goodAtClimb: number;
  goodAtDefense: number;
  goodAtWorkingWithAlliance: number;
  goodAtDriving: number;
  goodAtAuto: number;
  removedAlgaeFromReef: number;
  comments: string;
  tbaAutoLine: null;
  tbaEndgamePark: null;
  tbaEndgameShallow: null;
  tbaEndgameDeep: null;
  autoCrossedRSL: number;
  autoCoralABL1: number;
  autoCoralAL2: number;
  autoCoralAL3: number;
  autoCoralAL4: number;
  autoCoralBL2: number;
  autoCoralBL3: number;
  autoCoralBL4: number;
  autoCoralCDL1: number;
  autoCoralCL2: number;
  autoCoralCL3: number;
  autoCoralCL4: number;
  autoCoralDL2: number;
  autoCoralDL3: number;
  autoCoralDL4: number;
  autoCoralEFL1: number;
  autoCoralEL2: number;
  autoCoralEL3: number;
  autoCoralEL4: number;
  autoCoralFL2: number;
  autoCoralFL3: number;
  autoCoralFL4: number;
  autoCoralGHL1: number;
  autoCoralGL2: number;
  autoCoralGL3: number;
  autoCoralGL4: number;
  autoCoralHL2: number;
  autoCoralHL3: number;
  autoCoralHL4: number;
  autoCoralIJL1: number;
  autoCoralIL2: number;
  autoCoralIL3: number;
  autoCoralIL4: number;
  autoCoralJL2: number;
  autoCoralJL3: number;
  autoCoralJL4: number;
  autoCoralKLL1: number;
  autoCoralKL2: number;
  autoCoralKL3: number;
  autoCoralKL4: number;
  autoCoralLL2: number;
  autoCoralLL3: number;
  autoCoralLL4: number;
  autoProcessor: number;
  autoNet: number;
  teleopL1: number;
  teleopL2: number;
  teleopL3: number;
  teleopL4: number;
  teleopProcessor: number;
  teleopNet: number;
  teleopPark: number;
  teleopAttemptedShallow: number;
  teleopAttemptedDeep: number;
  teleopSuccessfulShallow: number;
  teleopSuccessfulDeep: number;
  dataConfidence: null;
}

export default ISAScoutForm;